requirements.txt
pillow
requests
twitter
